Refactor typedefinisjoner + CallOption
