+++ src/components/_common/copyLink/copyLink.tsx
@@ -3,12 +3,10 @@ import {translator} from 'translations';
 import {classNames} from '../../../utils/classnames';
 import {StaticImage} from '../image/StaticImage';
 import {usePageConfig} from 'store/hooks/usePageConfig';
-import {logAmplitudeEvent} from '../../../utils/amplitude';
+import {analyticsEvents,logAmplitudeEvent} from '../../../utils/amplitude';
 
 import linkIcon from '/public/gfx/link.svg';
-
 import style from './copyLink.module.scss';
-import {analyticsEvents} from "../../../types/analyticsTaxonomy";
 
 type CopyLinkProps = {
     anchor: string;
