--- src/types/analyticsTaxonomy.ts
@@ -1,14 +1,0 @@
-export enum analyticsEvents {
-    NAVIGATION = 'navigere',
-    COPY_LINK = 'kopier-lenke',
-    CHAT_OPEN = 'chat-åpnet',
-    CALL = 'ring-oss',
-}
-
-export enum analyticsContent {
-    TABLE_OF_CONTENTS = 'innholdsmeny',
-    CONTENT_SECTION = 'innholdsseksjon',
-    PRODUCT_LIST = 'varehylle',
-    OTHER_SUPPLIERS = 'andre som kan hjelpe',
-    OTHER_CHANNELS = 'kontaktmodul'
-}
